refactor(components): refactored accordion
fix #190
