+++ packages/components/src/base/Accordion/index.tsx
@@ -1,4 +1,1 @@
-export { Accordion } from './accordion';
-export { AccordionActions } from './accordionactions';
-export { AccordionDetails } from './accordiondetails';
-export { AccordionSummary } from './accordionsummary';
+export { Accordion } from './Accordion';
