+++ packages/components/src/base/AccordionActions/index.tsx
@@ -1,0 +1,1 @@
+export { AccordionActions } from './AccordionActions';
