+++ packages/components/src/base/AccordionDetails/AccordionDetails.tsx
