+++ packages/components/src/base/AccordionDetails/index.tsx
@@ -1,0 +1,1 @@
+export { AccordionDetails } from './AccordionDetails';
