+++ packages/components/src/base/AccordionSummary/index.tsx
@@ -1,0 +1,1 @@
+export { AccordionSummary } from './AccordionSummary';
